VOT_web/simple_login/requirements.txt
flask
mysql-connector-python
python-dotenv
werkzeug

